rename docs folder to "docsParts"
